requirements.txt
pandas
openai
python-dotenv
snowflake-connector-python
